refactor(commands): extract bug/clone tests to sibling file
